cmd/govim: use visual positions for diagnostics
When displaying diagnostics, the position must be visible to the user
within Vim. Some diagnostics have start/end positions that are beyond
the visible content, i.e. they are after the final \n. The user, however
will not be able to see/understand such a position. So we adjust such
positions to be before the final \n.