cmd/govim: add config for defining the gopls environment
gopls can be configured with an environment, an environment which is
passed to go/packages. For example, setting GOOS and GOARCH, or setting
GOFLAGS=-tags=other to set a build tag.

Therefore we define a govim config key, "GoplsEnv", for setting this
environment.

For example:

  govim#config#Set("GoplsEnv", {"GOFLAGS": "-mod=readonly"})

As part of this add a test that verifies the behaviour of the above
setting. Setting of build tags does not current work; that is blocked on
https://github.com/golang/go/issues/35548 and will be tested in a follow
up PR.

Fixes #555